refactor: move update_catalog_item to catalog_tools
Move UpdateCatalogItemParams and update_catalog_item from catalog_optimization.py
into catalog_tools.py where all other catalog CRUD operations live. Update all
import sites (tool_utils.py, tools/__init__.py) to source from catalog_tools.
Relocate the three update_catalog_item tests from test_catalog_optimization.py
to test_catalog_tools.py. Tool name and behavior unchanged.
